fix: disable problematic SauceDemo login users in CI
The SauceDemo test suite uses several different test users.
The problem, error, performance, and visual users are intermittently causing test failures in CI.
These users are not critical for the core functionality being tested, so they are being disabled in CI to improve the stability of the test suite.
They are disabled by commenting out the corresponding environment variables in the `selenium-tests.yaml` workflow file.