fix: change ShoppingCartPage.GetCartItems return type to List
This commit changes the return type of the `GetCartItems` method in the `ShoppingCartPage` class from `IEnumerable<CartItemComponent>` to `List<CartItemComponent>`. This change simplifies the usage of the method in the `SauceDemoTests.Cart.cs` test file, removing the need to convert the result to a list using `.ToList()`. The method documentation is also updated to reflect the change.